Add missing length calculation after interface refactor

+++ include/cryptography_interface.h
@@ -38,8 +38,8 @@ typedef struct
     int32_t (*cryptography_decrypt)(void);
     int32_t (*cryptography_authenticate)(void);
     int32_t (*cryptography_validate_authentication)(void);
-    int32_t (*cryptography_aead_encrypt)(uint8_t* data_out, uint32_t len_data_out,
-                                         uint8_t* data_in, uint32_t len_data_in,
+    int32_t (*cryptography_aead_encrypt)(uint8_t* data_out, size_t len_data_out,
+                                         uint8_t* data_in, size_t len_data_in,
                                          uint8_t* key, uint32_t len_key,
                                          SecurityAssociation_t* sa_ptr,
                                          uint8_t* iv, uint32_t iv_len,
@@ -47,8 +47,8 @@ typedef struct
                                          uint8_t* aad, uint32_t aad_len,
                                          uint8_t encrypt_bool, uint8_t authenticate_bool,
                                          uint8_t aad_bool);
-    int32_t (*cryptography_aead_decrypt)(uint8_t* data_out, uint32_t len_data_out,
-                                         uint8_t* data_in, uint32_t len_data_in,
+    int32_t (*cryptography_aead_decrypt)(uint8_t* data_out, size_t len_data_out,
+                                         uint8_t* data_in, size_t len_data_in,
                                          uint8_t* key, uint32_t len_key,
                                          SecurityAssociation_t* sa_ptr,
                                          uint8_t* iv, uint32_t iv_len,
