+++ src/src_main/crypto_tm.c
@@ -217,22 +217,22 @@ int32_t Crypto_TM_ApplySecurity(uint8_t *ingest, int *len_ingest)
 #endif
 
         status = cryptography_if->cryptography_aead_encrypt(&(ingest[pdu_loc]), // ciphertext output
-                                                   pdu_len,            // length of data
-                                                   &(tempTM[pdu_loc]), // plaintext input
-                                                   pdu_len,             // in data length
-                                                   NULL, // Key is mapped via SA
-                                                   KEY_SIZE,
-                                                   sa_ptr,
-                                                   sa_ptr->iv,
-                                                   sa_ptr->shivf_len,
-                                                   &(ingest[mac_loc]),
-                                                   MAC_SIZE,
-                                                   &(aad[0]), // AAD Input location
-                                                   sa_ptr->abm_len, // AAD is size of ABM in this case
-                                                   CRYPTO_TRUE, // Encrypt
-                                                   CRYPTO_FALSE, // Authenticate // TODO -- Set to SA value, manually setting to false here so existing tests pass. Existing data was generated with authenticate then encrypt, when it should have been encrypt then authenticate.
-                                                   CRYPTO_TRUE // Use AAD
-                                                   );
+                                                           (size_t)pdu_len,            // length of data
+                                                           &(tempTM[pdu_loc]), // plaintext input
+                                                           (size_t)pdu_len,             // in data length
+                                                           NULL, // Key is mapped via SA
+                                                           KEY_SIZE,
+                                                           sa_ptr,
+                                                           sa_ptr->iv,
+                                                           sa_ptr->shivf_len,
+                                                           &(ingest[mac_loc]),
+                                                           MAC_SIZE,
+                                                           &(aad[0]), // AAD Input location
+                                                           sa_ptr->abm_len, // AAD is size of ABM in this case
+                                                           CRYPTO_TRUE, // Encrypt
+                                                           CRYPTO_FALSE, // Authenticate // TODO -- Set to SA value, manually setting to false here so existing tests pass. Existing data was generated with authenticate then encrypt, when it should have been encrypt then authenticate.
+                                                           CRYPTO_TRUE // Use AAD
+                                                           );
 
 
         // Update OCF
